+++ apps/docs/components/icons/bold/index.ts
@@ -16,3 +16,4 @@ export * from "./hash";
 export * from "./more-square";
 export * from "./play";
 export * from "./pause";
+export * from "./cli";