+++ packages/components/checkbox/src/use-checkbox-group.ts
@@ -1,15 +1,17 @@
 import type {CheckboxGroupSlots, SlotsToClasses} from "@nextui-org/theme";
 import type {AriaCheckboxGroupProps} from "@react-types/checkbox";
 import type {Orientation} from "@react-types/shared";
-import type {HTMLNextUIProps, PropGetter} from "@nextui-org/system";
 import type {ReactRef} from "@nextui-org/react-utils";
+import type {CheckboxGroupProps} from "@react-types/checkbox";
+import type {HTMLNextUIProps, PropGetter} from "@nextui-org/system";
 
+import {useProviderContext} from "@nextui-org/system";
 import {useCallback, useMemo} from "react";
-import {mergeProps} from "@react-aria/utils";
+import {chain, mergeProps} from "@react-aria/utils";
 import {checkboxGroup} from "@nextui-org/theme";
 import {useCheckboxGroup as useReactAriaCheckboxGroup} from "@react-aria/checkbox";
 import {CheckboxGroupState, useCheckboxGroupState} from "@react-stately/checkbox";
-import {useDOMRef} from "@nextui-org/react-utils";
+import {filterDOMProps, useDOMRef} from "@nextui-org/react-utils";
 import {clsx, safeAriaLabel} from "@nextui-org/shared-utils";
 
 import {CheckboxProps} from "./index";
@@ -46,7 +48,7 @@ interface Props extends HTMLNextUIProps<"div"> {
   onValueChange?: AriaCheckboxGroupProps["onChange"];
 }
 
-export type UseCheckboxGroupProps = Props &
+export type UseCheckboxGroupProps = Omit<Props, "onChange"> &
   AriaCheckboxGroupProps &
   Partial<
     Pick<
@@ -64,9 +66,12 @@ export type ContextType = {
   lineThrough?: CheckboxProps["lineThrough"];
   isDisabled?: CheckboxProps["isDisabled"];
   disableAnimation?: CheckboxProps["disableAnimation"];
+  validationBehavior?: CheckboxProps["validationBehavior"];
 };
 
 export function useCheckboxGroup(props: UseCheckboxGroupProps) {
+  const globalContext = useProviderContext();
+
   const {
     as,
     ref,
@@ -77,14 +82,15 @@ export function useCheckboxGroup(props: UseCheckboxGroupProps) {
     value,
     name,
     defaultValue,
+    isInvalid: isInvalidProp,
+    validationState,
     size = "md",
     color = "primary",
     orientation = "vertical",
     lineThrough = false,
     isDisabled = false,
-    disableAnimation = false,
-    validationState,
-    isInvalid = validationState === "invalid",
+    validationBehavior = globalContext?.validationBehavior ?? "aria",
+    disableAnimation = globalContext?.disableAnimation ?? false,
     isReadOnly,
     isRequired,
     onValueChange,
@@ -95,43 +101,50 @@ export function useCheckboxGroup(props: UseCheckboxGroupProps) {
   } = props;
 
   const Component = as || "div";
+  const shouldFilterDOMProps = typeof Component === "string";
 
   const domRef = useDOMRef(ref);
 
-  const checkboxGroupProps = useMemo(
-    () => ({
+  const checkboxGroupProps = useMemo<CheckboxGroupProps>(() => {
+    return {
+      ...otherProps,
       value,
       name,
       "aria-label": safeAriaLabel(otherProps["aria-label"], label),
       defaultValue,
       isRequired,
-      isInvalid,
       isReadOnly,
       orientation,
-      onChange: onValueChange,
-      ...otherProps,
-    }),
-    [
-      value,
-      name,
-      label,
-      defaultValue,
-      isRequired,
-      isReadOnly,
-      isInvalid,
-      orientation,
-      onValueChange,
-      otherProps["aria-label"],
-      otherProps,
-    ],
-  );
+      validationBehavior,
+      isInvalid: validationState === "invalid" || isInvalidProp,
+      onChange: chain(props.onChange, onValueChange),
+    };
+  }, [
+    value,
+    name,
+    label,
+    defaultValue,
+    isRequired,
+    isReadOnly,
+    orientation,
+    onValueChange,
+    isInvalidProp,
+    validationState,
+    validationBehavior,
+    otherProps["aria-label"],
+    otherProps,
+  ]);
 
   const groupState = useCheckboxGroupState(checkboxGroupProps);
 
-  const {labelProps, groupProps, descriptionProps, errorMessageProps} = useReactAriaCheckboxGroup(
-    checkboxGroupProps,
-    groupState,
-  );
+  const {
+    labelProps,
+    groupProps,
+    descriptionProps,
+    errorMessageProps,
+    validationErrors,
+    validationDetails,
+  } = useReactAriaCheckboxGroup(checkboxGroupProps, groupState);
 
   const context = useMemo<ContextType>(
     () => ({
@@ -139,9 +152,10 @@ export function useCheckboxGroup(props: UseCheckboxGroupProps) {
       color,
       radius,
       lineThrough,
-      isInvalid,
+      isInvalid: groupState.isInvalid,
       isDisabled,
       disableAnimation,
+      validationBehavior,
       groupState,
     }),
     [
@@ -151,18 +165,18 @@ export function useCheckboxGroup(props: UseCheckboxGroupProps) {
       lineThrough,
       isDisabled,
       disableAnimation,
-      isInvalid,
-      groupState?.value,
-      groupState?.isDisabled,
-      groupState?.isReadOnly,
-      groupState?.isInvalid,
-      groupState?.isSelected,
+      validationBehavior,
+      groupState.value,
+      groupState.isDisabled,
+      groupState.isReadOnly,
+      groupState.isInvalid,
+      groupState.isSelected,
     ],
   );
 
   const slots = useMemo(
-    () => checkboxGroup({isRequired, isInvalid, disableAnimation}),
-    [isRequired, isInvalid, disableAnimation],
+    () => checkboxGroup({isRequired, isInvalid: groupState.isInvalid, disableAnimation}),
+    [isRequired, groupState.isInvalid, , disableAnimation],
   );
 
   const baseStyles = clsx(classNames?.base, className);
@@ -171,7 +185,12 @@ export function useCheckboxGroup(props: UseCheckboxGroupProps) {
     return {
       ref: domRef,
       className: slots.base({class: baseStyles}),
-      ...mergeProps(groupProps, otherProps),
+      ...mergeProps(
+        groupProps,
+        filterDOMProps(otherProps, {
+          enabled: shouldFilterDOMProps,
+        }),
+      ),
     };
   }, [slots, domRef, baseStyles, groupProps, otherProps]);
 
@@ -218,7 +237,11 @@ export function useCheckboxGroup(props: UseCheckboxGroupProps) {
     label,
     context,
     description,
-    errorMessage,
+    isInvalid: groupState.isInvalid,
+    errorMessage:
+      typeof errorMessage === "function"
+        ? errorMessage({isInvalid: groupState.isInvalid, validationErrors, validationDetails})
+        : errorMessage || validationErrors?.join(" "),
     getGroupProps,
     getLabelProps,
     getWrapperProps,