+++ packages/components/dropdown/stories/dropdown.stories.tsx
@@ -84,6 +84,11 @@ export default {
         type: "boolean",
       },
     },
+    isDisabled: {
+      control: {
+        type: "boolean",
+      },
+    },
     disableAnimation: {
       control: {
         type: "boolean",
@@ -109,8 +114,8 @@ const defaultProps = {
   ...dropdown.defaultVariants,
   placement: "bottom",
   offset: 7,
+  isDisabled: false,
   defaultOpen: false,
-  disableAnimation: false,
 };
 
 const items = [
@@ -137,7 +142,7 @@ const Template = ({color, variant, ...args}: DropdownProps & DropdownMenuProps) 
     <DropdownTrigger>
       <Button>Trigger</Button>
     </DropdownTrigger>
-    <DropdownMenu aria-label="Actions" color={color} variant={variant} onAction={alert}>
+    <DropdownMenu aria-label="Actions" color={color} variant={variant}>
       <DropdownItem key="new">New file</DropdownItem>
       <DropdownItem key="copy">Copy link</DropdownItem>
       <DropdownItem key="edit">Edit file</DropdownItem>
@@ -146,6 +151,41 @@ const Template = ({color, variant, ...args}: DropdownProps & DropdownMenuProps) 
       </DropdownItem>
     </DropdownMenu>
   </Dropdown>
+);
+
+const ItemCloseOnSelectTemplate = ({
+  color,
+  variant,
+  ...args
+}: DropdownProps & DropdownMenuProps) => (
+  <>
+    <Dropdown {...args}>
+      <DropdownTrigger>
+        <Button>Trigger</Button>
+      </DropdownTrigger>
+      <DropdownMenu aria-label="Actions" color={color} variant={variant}>
+        <DropdownItem key="new">New file</DropdownItem>
+        <DropdownItem key="copy">Copy link</DropdownItem>
+        <DropdownItem key="edit">Edit file</DropdownItem>
+        <DropdownItem key="delete" className="text-danger" closeOnSelect={false} color="danger">
+          Delete file
+        </DropdownItem>
+      </DropdownMenu>
+    </Dropdown>
+    <Dropdown {...args}>
+      <DropdownTrigger>
+        <Button>Trigger</Button>
+      </DropdownTrigger>
+      <DropdownMenu aria-label="Actions" closeOnSelect={false} color={color} variant={variant}>
+        <DropdownItem key="new">New file</DropdownItem>
+        <DropdownItem key="copy">Copy link</DropdownItem>
+        <DropdownItem key="edit">Edit file</DropdownItem>
+        <DropdownItem key="delete" className="text-danger" color="danger">
+          Delete file
+        </DropdownItem>
+      </DropdownMenu>
+    </Dropdown>
+  </>
 );
 
 const DynamicTemplate = ({color, variant, ...args}: DropdownProps & DropdownMenuProps) => (
@@ -569,6 +609,24 @@ const CustomTriggerTemplate = ({variant, ...args}) => {
   );
 };
 
+const CustomHTMLTrigger = ({variant, ...args}) => {
+  return (
+    <Dropdown {...args}>
+      <DropdownTrigger>
+        <span className="flex items-center gap-2">Profile</span>
+      </DropdownTrigger>
+      <DropdownMenu aria-label="Actions" variant={variant}>
+        <DropdownItem key="new">New file</DropdownItem>
+        <DropdownItem key="copy">Copy link</DropdownItem>
+        <DropdownItem key="edit">Edit file</DropdownItem>
+        <DropdownItem key="delete" className="text-danger" color="danger">
+          Delete file
+        </DropdownItem>
+      </DropdownMenu>
+    </Dropdown>
+  );
+};
+
 export const Default = {
   render: Template,
 
@@ -607,6 +665,15 @@ export const DisabledKeys = {
 
   args: {
     ...defaultProps,
+  },
+};
+
+export const DisabledTrigger = {
+  render: Template,
+
+  args: {
+    ...defaultProps,
+    isDisabled: true,
   },
 };
 
@@ -686,6 +753,16 @@ export const WithCustomTrigger = {
   },
 };
 
+export const WithCustomHTMLTrigger = {
+  render: CustomHTMLTrigger,
+
+  args: {
+    ...defaultProps,
+    variant: "flat",
+    offset: 14,
+  },
+};
+
 export const DisableAnimation = {
   render: WithStartContentTemplate,
 
@@ -695,5 +772,13 @@ export const DisableAnimation = {
     variant: "flat",
     color: "secondary",
     disableAnimation: true,
+  },
+};
+
+export const ItemCloseOnSelect = {
+  render: ItemCloseOnSelectTemplate,
+
+  args: {
+    ...defaultProps,
   },
 };