+++ packages/components/pagination/stories/pagination.stories.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {Meta} from "@storybook/react";
 import {button, pagination} from "@nextui-org/theme";
-import {cn} from "@nextui-org/system";
+import {cn} from "@nextui-org/theme";
 import {ChevronIcon} from "@nextui-org/shared-icons";
 
 import {Pagination, PaginationItemRenderProps, PaginationItemType, usePagination} from "../src";