+++ packages/core/system-rsc/src/index.ts
@@ -1,3 +1,31 @@
-export * from "./types";
-export * from "./utils";
-export * from "./extend-variants";
+export type {
+  As,
+  DOMElement,
+  DOMElements,
+  CapitalizedDOMElements,
+  DOMAttributes,
+  OmitCommonProps,
+  RightJoinProps,
+  MergeWithAs,
+  InternalForwardRefRenderFunction,
+  PropsOf,
+  Merge,
+  HTMLNextUIProps,
+  PropGetter,
+  SharedSelection,
+} from "./types";
+
+export {
+  forwardRef,
+  toIterator,
+  mapPropsVariants,
+  mapPropsVariantsWithCommon,
+  isNextUIEl,
+} from "./utils";
+
+export {extendVariants} from "./extend-variants";
+export type {
+  ExtendVariantProps,
+  ExtendVariantWithSlotsProps,
+  ExtendVariants,
+} from "./extend-variants";